refactor: extract runAutoSyncIfEnabled into lib/auto-sync-runner.ts factory
Move the 450-line runAutoSyncIfEnabled function from content.ts into
a new DI factory (createAutoSyncRunner). Mutable state
(autoSyncTriggered, autoSyncCancelled) is passed as a shared object
reference to avoid temporal dead zone issues.

content.ts: 1436→1153 lines (-283)
